PSA eBPF: Adding AssignRandomValue command
Adds the AssignRandomValue command. To get a
random value in the PSA architecture you need
to instatiate a random value generator in the
a_declarations section. This is achieved by
declaring it as a state variable. To get a random
value only the variable and the name of the random
number genrator need to be supplied to the
AssignRandomValue constructor. This will end up to
call the read method of the rng.